Implement the adaptor
Created a new class called ArticleAdaptor which uses a viewholder to
update the recyclerView items with the data from the cursor.

Created constants for the article columns and indeces to make it easier
to access the data.

Moved the typefaces into the adaptor.

Currently compiling but not showing data from the cursor.