Fix certain articles not loading images
The link to the article was used as a default value for article images
during setup but now it's changed to use the thumbnail (which is too low
resolution normally but is better than no image).

Also moved values for unique rows to only be added to the content values
if the row doesn't already exist.