Develop detail page layout
Created a bar to show the article title and subtitle on. Also created a
bulletpoint_item file to develop the bulletpoint. These files could then
be included into the main xml to save code duplication.

The whole block was wrapped in a scrollview so the entire section could
be scrolled.