Write BulletPointWizard and get first bulletpoints
BulletPointWizard has been developed so that it can take in the body of
the article and split the article into sentences and then score these
sentences. The first and last sentences, and then the following lowest
are chosen and returned.

This has changed from how it was planned in the design document due to
a majority of the paragraphs in an article being one sentence long, so
splitting it by paragraphs was redundant.

Quick but ugly changes were also made to show it up on the UI to get
some quick visual feedback.