Fix issue where not all bulletpoints would load
Sometimes, if there are few sentences, only some of the bulletpoints
were being chosen. This was overcome by setting default values before
the loop so it would always have a value.

There was also the possibility of overlap due to the <= and >= signs so
they were replaced to ensure each section is picking from its own area.