Fix bulletpoint generator
Generator was sometimes selecting the same sentence if the amount of
sentences available was 4 or below. This was fixed by not overwriting
the defaults unless the sentence count was above 4.

There was also an issue where the extra detail being generated didn't
necessarily include the sentence that was being clicked on. This was
because the extra detail wasn't updated along with the sentence.

Also refactored the default generation into its own method.