Fix UI flicker when articles were being loaded
Each article was creating a callback to get the html data which was
triggering onLoadFinished multiple times when the app was first launched
and when it was refreshed.

The data collection was moved onto the same background thread as the
rest of the data fetcher, and then the values were uploaded with a bulk
insert operation.